refactor: restructure project directory layout
